+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v0.15.2
 - Added SeqNo to `WriteResult`.
 - Changed signature of the `TopicClient.DropTopic` method.